#115 - Refactoring code (not finished)
